+++ ios/react-native-navigation-sdk/NavModule.mm
@@ -15,9 +15,11 @@
  */
 
 #import "NavModule.h"
+#import <GoogleMaps/GoogleMaps.h>
 #import "NavAutoModule.h"
 #import "NavViewModule.h"
 #import "ObjectTranslationUtil.h"
+#import "SdkVersion.h"
 
 using namespace JS::NativeNavModule;
 
@@ -104,10 +106,14 @@ RCT_EXPORT_MODULE(NavModule);
   _session.started = YES;
 
   if (self->_session.navigator) {
+    // Remove any existing listener first to prevent duplicates
+    // in case initializeSession is called multiple times.
+    [self->_session.navigator removeListener:self];
     [self->_session.navigator addListener:self];
     self->_session.navigator.stopGuidanceAtArrival = NO;
   }
 
+  [self->_session.roadSnappedLocationProvider removeListener:self];
   [self->_session.roadSnappedLocationProvider addListener:self];
 
   NavViewModule *navViewModule = [NavViewModule sharedInstance];
@@ -277,6 +283,7 @@ RCT_EXPORT_MODULE(NavModule);
       [self->_session.roadSnappedLocationProvider removeListener:self];
     }
 
+    self.enableUpdateInfo = NO;
     self->_session.started = NO;
     self->_session = nil;
 
@@ -415,8 +422,18 @@ RCT_EXPORT_MODULE(NavModule);
       return;
     }
 
-    [navigator continueToNextDestination];
-    resolve(@(YES));
+    [navigator continueToNextDestinationWithCompletion:^(GMSNavigationWaypoint *_Nullable waypoint,
+                                                         GMSRouteStatus routeStatus) {
+      NSMutableDictionary *result = [[NSMutableDictionary alloc] init];
+      if (waypoint != nil) {
+        result[@"waypoint"] =
+            [ObjectTranslationUtil transformNavigationWaypointToDictionary:waypoint];
+      } else {
+        result[@"waypoint"] = [NSNull null];
+      }
+      result[@"routeStatus"] = [NavModule routeStatusToString:routeStatus];
+      resolve(result);
+    }];
   });
 }